requirements.txt
safety==1.8.5
bandit==1.6.2
pytest==5.2.1
pytest-cov==2.5.1
pytest-django==3.7.0
pytest-sugar==0.9.2
django-cors-headers
django-countries
python-jose==3.0.1
auth0-python==3.9.1
pytest-xdist==1.30.0
google-cloud-storage==1.23.0
earthengine-api
openpyxl==3.0.3
